aider: Fix: updated code to correctly access the `KV` namespace from the `env` object in a Cloudflare Worker environment.
# Aider chat conversation:

USER: ✘ [ERROR] Error: KVNamespace is not a constructor

ASSISTANT: To fix the error `KVNamespace is not a constructor`, we need to update the code to use the correct way to interact with Cloudflare KV in a Cloudflare Worker environment.

Here are the *SEARCH/REPLACE* blocks to fix this issue:

src/pwall.js
```javascript
<<<<<<< SEARCH
const { KVNamespace } = require('@cloudflare/kv-asset-handler');
=======
>>>>>>> REPLACE
```

src/pwall.js
```javascript
<<<<<<< SEARCH
        const kv = new KVNamespace({ binding: 'KV', apiToken: env.CLOUDFLARE_API_TOKEN });
=======
        const kv = env.KV;
>>>>>>> REPLACE
```

This will ensure that the `KV` namespace is correctly accessed from the `env` object provided by the Cloudflare Worker environment.