aider: Added console.log statements for better visibility of timing information.
# Aider chat conversation:

USER: I'm not seeing the output for console.time or console.timeEnd, why is that?

ASSISTANT: The `console.time` and `console.timeEnd` methods log their output to the console. If you are not seeing the output, it might be because the environment where the code is running does not support these methods or the logs are not being displayed.

To ensure that the timing information is logged, we can use `console.log` to explicitly log the start and end times. Here are the *SEARCH/REPLACE* blocks to add these logs:

src/index.js
```javascript
<<<<<<< SEARCH
    const voltage = env.voltage;
=======
    console.log('handleFetch started');
    const voltage = env.voltage;
    console.log('Fetch current and previous day keys started');
>>>>>>> REPLACE
```

src/index.js
```javascript
<<<<<<< SEARCH
    if ((!currentDayKeys || !currentDayKeys.keys || currentDayKeys.keys.length === 0) &&
=======
    console.log('Fetch current and previous day keys ended');
    if ((!currentDayKeys || !currentDayKeys.keys || currentDayKeys.keys.length === 0) &&
>>>>>>> REPLACE
```

src/index.js
```javascript
<<<<<<< SEARCH
    for (const key of allKeys) {
=======
    console.log('Process all keys started');
    for (const key of allKeys) {
>>>>>>> REPLACE
```

src/index.js
```javascript
<<<<<<< SEARCH
    const htmlTemplate = `
=======
    console.log('Process all keys ended');
    console.log('Generate HTML template started');
    const htmlTemplate = `
>>>>>>> REPLACE
```

src/index.js
```javascript
<<<<<<< SEARCH
    return new Response(htmlTemplate, { status: 200, headers: { 'Content-Type': 'text/html' } });
=======
    console.log('Generate HTML template ended');
    console.log('handleFetch ended');
    return new Response(htmlTemplate, { status: 200, headers: { 'Content-Type': 'text/html' } });
>>>>>>> REPLACE
```