aider: Declare `allKeysValues` before it is used in the `handleFetch` function.
# Aider chat conversation:

USER: now `allKeysValues` is not defined on line 80

ASSISTANT: To fix the issue of `allKeysValues` not being defined on line 80, we need to declare it before it is used. Here are the necessary changes:

src/index.js
```javascript
<<<<<<< SEARCH
        const fetchKeysEndTime = Date.now();
=======
        const allKeysValues = {};
        const fetchKeysEndTime = Date.now();
>>>>>>> REPLACE
```

This will ensure that `allKeysValues` is declared before it is used in the `handleFetch` function.