aider: Modified handleJson function to include the number of calls in the HTML response and added a link to download the JSON file.
# Aider chat conversation:

USER: i just want to see the getCalls in the webpage.  but also create a json file for download.  is that possible?

ASSISTANT: To show the `getCalls` on the webpage and also create a JSON file for download, we need to:

1. Modify the `handleJson` function to include the number of calls in the HTML response.
2. Add a link to download the JSON file.

Here are the *SEARCH/REPLACE* blocks:

src/index.js
```javascript
<<<<<<< SEARCH
        const htmlTemplate = `
=======
        const jsonContent = JSON.stringify({ getCalls, data: allKeysValues }, null, 2);
        const htmlTemplate = `
>>>>>>> REPLACE
```

src/index.js
```javascript
<<<<<<< SEARCH
            <pre>${jsonContent}</pre>
=======
            <a href="data:text/json;charset=utf-8,${encodeURIComponent(jsonContent)}" download="voltage_data.json">Download JSON</a>
            <pre>${JSON.stringify({ getCalls, data: allKeysValues }, null, 2)}</pre>
>>>>>>> REPLACE
```