aider: Initialize `getCalls` at the beginning of `handleJson` to ensure it is always defined, even in case of errors.
# Aider chat conversation:

USER: in `handleJson` the error template usees getCalls but if the error is in the promise, getCalls won't be created, correct?

ASSISTANT: To ensure that `getCalls` is always defined, we can initialize it at the beginning of the `handleJson` function. This way, even if an error occurs, `getCalls` will be available in the error template.

Here are the *SEARCH/REPLACE* blocks:

src/index.js
```javascript
<<<<<<< SEARCH
    try {
=======
    let getCalls = 0;
    try {
>>>>>>> REPLACE
```

src/index.js
```javascript
<<<<<<< SEARCH
        const getCalls = getCallsResults.reduce((acc, curr) => acc + curr, 0);
=======
        getCalls = getCallsResults.reduce((acc, curr) => acc + curr, 0);
>>>>>>> REPLACE
```